requirements.txt
psycopg2-binary==2.9.9
selenium==4.18.1
python-dotenv==1.0.1 